Modification on the way edges are stored
Instead of storing edges using their position inside the global tree, we store
their BFS index locally in the subtree to which the tree is linked along with
the encoding index of this tree. In other words, we store a tuple (Int, Int),
where the first Int is the parsing index, the second the position in BFS
of the edge, locally,inside the subtree referenced by the parsing index.