ToolBox with working Universe incl + refact. sbt/
- "Mutual" dependencies between ToolBox and TreeRecomposer OK.
- Removal in decompression of the type DecompressTree, which was a trouble
because dependent of Universe also. Moreover, since we wiil probably not need
it for long (we don't take care of names, types, etc), this would have
been changed later.

- Refactoring of the sbt/ folder : two projects, one for the SBT plugin and
another one for tests. Since it is apparently not possible to design an
automatic inclusion of a sbt plugin into subprojects.
- The automatic inclusion of the NSC plugin doesn't work. Yet.